devpi: Allow to build it
Add a simple "installPhase" so that it can be built using nix-build. A more
accurate way would probably be to use one of the "buildEnv" like functions, but
the sole purpose for the moment is to be able to build it to see if it works out
on remote systems as well.